+++ app/assets/javascripts/application.js
@@ -28,3 +28,21 @@
 //
 //= require rails-ujs
 //= require_tree .
+
+// If we've got test keys enabled on the site, then we'll be altering the colours to give some sense of a staging
+// environment. We can turn this off to test look and feel using the button below.
+$(document).ready(function initStyleToggler() {
+  if (!$('body').hasClass("api-test-keys")) {
+    return;
+  }
+
+  var button = $("<input>").attr({
+    value: "Toggle Styles",
+    title: "Stripe test keys present, this makes it clear you're in a staging environment",
+    class: "btn btn-api-toggle-styles",
+  });
+  $("body").append(button);
+  button.on("click", function() {
+    $('body').toggleClass("api-test-keys");
+  });
+});