+++ app/controllers/application_controller.rb
@@ -17,8 +17,6 @@
 class ApplicationController < ActionController::Base
   layout "conzealand"
 
-  protect_from_forgery with: :exception
-
   protected
 
   def lookup_purchase