+++ config/environments/production.rb
@@ -30,12 +30,12 @@ Rails.application.configure do
   config.consider_all_requests_local       = false
   config.action_controller.perform_caching = true
 
-  # Attempt to read encrypted secrets from `config/secrets.yml.enc`.
-  # Requires an encryption key in `ENV["RAILS_MASTER_KEY"]` or
-  # `config/secrets.yml.key`.
-  config.read_encrypted_secrets = true
+  # Ensures that a master key has been made available in either ENV["RAILS_MASTER_KEY"]
+  # or in config/master.key. This key is used to decrypt credentials (and other encrypted files).
+  # config.require_master_key = true
 
-  # This serves up files from public/assets on Production.
+  # Disable serving static files from the `/public` folder by default since
+  # Apache or NGINX already handles this.
   config.public_file_server.enabled = true
 
   # Compress JavaScripts and CSS.
@@ -53,6 +53,9 @@ Rails.application.configure do
   # Specifies the header that your server uses for sending files.
   # config.action_dispatch.x_sendfile_header = 'X-Sendfile' # for Apache
   # config.action_dispatch.x_sendfile_header = 'X-Accel-Redirect' # for NGINX
+
+  # Store uploaded files on the local file system (see config/storage.yml for options)
+  config.active_storage.service = :local
 
   # Mount Action Cable outside main process or domain
   # config.action_cable.mount_path = nil
@@ -75,8 +78,12 @@ Rails.application.configure do
   # Use a real queuing backend for Active Job (and separate queues per environment)
   # config.active_job.queue_adapter     = :resque
   # config.active_job.queue_name_prefix = "conzealand_#{Rails.env}"
+
   config.action_mailer.perform_caching = false
-  config.action_mailer.delivery_method = :smtp
+
+  # Ignore bad email addresses and do not raise email delivery errors.
+  # Set this to true and configure the email server for immediate delivery to raise delivery errors.
+  # config.action_mailer.raise_delivery_errors = false
 
   # Enable locale fallbacks for I18n (makes lookups for any locale fall back to
   # the I18n.default_locale when a translation cannot be found).
@@ -109,11 +116,9 @@ Rails.application.configure do
     protocol: "https",
   }
 
-  # Don't fill logs with colour codes
-  config.colorize_logging = false
-
   # We're using docker to handle logging, so we're setting this to standard out.
   # If this doesn't work for your usecase, please raise an issue.
   # See https://success.docker.com/article/logging-best-practices
   Rails.logger = ActiveSupport::Logger.new(STDOUT)
+  config.colorize_logging = false
 end