+++ config/initializers/email.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# Copyright 2018 Matthew B. Gray
+# Copyright 2019 Matthew B. Gray
 # Copyright 2019 Steven C Hartley
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,17 +15,37 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# https://guides.rubyonrails.org/action_mailer_basics.html
-if !Rails.env.development?
-  Rails.application.config.action_mailer.tap do |action_mailer|
-    action_mailer.raise_delivery_errors = true
-    action_mailer.smtp_settings = {
-      address:              ENV["SMTP_SERVER"],
-      port:                 ENV["SMTP_PORT"],
-      user_name:            ENV["SMTP_USER_NAME"],
-      password:             ENV["SMTP_PASSWORD"],
-      authentication:       "plain",
-      enable_starttls_auto: true
+Rails.application.config.action_mailer.tap do |action_mailer|
+  if Rails.env.development?
+    # Advice from rails g devise:install
+    action_mailer.default_url_options = { host: "localhost", port: 3000 }
+
+    # mail catcher
+    action_mailer.delivery_method = :smtp
+    action_mailer.smtp_settings = { address: "0.0.0.0", port: 1025 }
+
+    # Mailer previews for rspec
+    # see https://stackoverflow.com/a/39204340/81271
+    # preview on http://localhost:3000/rails/mailers
+    action_mailer.preview_path = "#{Rails.root}/spec/mailers/previews"
+  elsif Rails.env.test?
+    action_mailer.default_url_options = {
+      host: "localhost"
     }
+  else
+    # Setup SMTP
+    # https://guides.rubyonrails.org/action_mailer_basics.html
+    action_mailer.delivery_method = :smtp
+    action_mailer.tap do |action_mailer|
+      action_mailer.raise_delivery_errors = true
+      action_mailer.smtp_settings = {
+        address:              ENV["SMTP_SERVER"],
+        port:                 ENV["SMTP_PORT"],
+        user_name:            ENV["SMTP_USER_NAME"],
+        password:             ENV["SMTP_PASSWORD"],
+        authentication:       "plain",
+        enable_starttls_auto: true
+      }
+    end
   end
 end