+++ config/routes.rb
@@ -16,6 +16,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+# For more information about routes, see http://guides.rubyonrails.org/routing.html
 Rails.application.routes.draw do
   root to: "landing#index"